refactor: move code to update state into separate file
