+++ lua/crates.lua
@@ -50,119 +50,14 @@ local M = {}
 
 
 
-local api = require("crates.api")
-local Version = api.Version
-local Dependency = api.Dependency
 local config = require("crates.config")
 local Config = config.Config
+local state = require("crates.state")
 local core = require("crates.core")
 local popup = require("crates.popup")
-local toml = require("crates.toml")
-local Crate = toml.Crate
-local diagnostic = require("crates.diagnostic")
 local util = require("crates.util")
 local ui = require("crates.ui")
 local Range = require("crates.types").Range
-
-local function reload_deps(crate, versions, version)
-   local function on_fetched(deps, cancelled)
-      if cancelled then return end
-
-      if deps then
-         version.deps = deps
-         for _, d in ipairs(deps) do
-            if d.opt and not version.features:get_feat(d.name) then
-               table.insert(version.features, {
-                  name = d.name,
-                  members = {},
-               })
-            end
-         end
-         version.features:sort()
-
-         for b, crates in pairs(core.crate_cache) do
-            for _, c in pairs(crates) do
-               if c.name == crate.name then
-                  local avoid_pre = core.cfg.avoid_prerelease and not c:vers_is_pre()
-                  local m, p, y = util.get_newest(versions, avoid_pre, c:vers_reqs())
-                  local match = m or p or y
-
-                  if c.vers and match == version and vim.api.nvim_buf_is_loaded(b) then
-                     local diagnostics = diagnostic.process_crate_deps(c, version, deps)
-                     ui.display_diagnostics(b, diagnostics)
-                  end
-               end
-            end
-         end
-      end
-   end
-
-   api.fetch_crate_deps(crate.name, version.num, on_fetched)
-end
-
-local function reload_crate(buf, crate)
-   local function on_fetched(versions, cancelled)
-      if cancelled then return end
-
-      if versions and versions[1] then
-         core.vers_cache[crate.name] = versions
-      end
-
-      for b, crates in pairs(core.crate_cache) do
-         for _, c in pairs(crates) do
-
-            if c.name == crate.name and vim.api.nvim_buf_is_loaded(b) then
-               local info = diagnostic.process_crate_versions(c, versions)
-               ui.display_crate_info(b, info)
-
-               if versions and versions[1] then
-                  reload_deps(c, versions, info.version or versions[1])
-               end
-            end
-         end
-      end
-   end
-
-   if core.cfg.loading_indicator then
-      ui.display_loading(buf, crate)
-   end
-
-   api.fetch_crate_versions(crate.name, on_fetched)
-end
-
-local function update(buf, reload)
-   if reload then
-      core.vers_cache = {}
-      api.cancel_jobs()
-   end
-
-   buf = buf or util.current_buf()
-   local sections, crates = toml.parse_crates(buf)
-   local cache, diagnostics = diagnostic.process_crates(sections, crates)
-
-   ui.clear(buf)
-   ui.display_diagnostics(buf, diagnostics)
-   for _, c in pairs(cache) do
-      local versions = core.vers_cache[c.name]
-
-      if not reload and versions then
-         local info = diagnostic.process_crate_versions(c, versions)
-         ui.display_crate_info(buf, info)
-
-         local version = info.version or versions[1]
-         if version.deps then
-            diagnostics = diagnostic.process_crate_deps(c, version, version.deps)
-            ui.display_diagnostics(buf, diagnostics)
-         else
-            reload_deps(c, versions, version)
-         end
-      else
-         reload_crate(buf, c)
-      end
-   end
-
-   core.crate_cache[buf] = cache
-end
 
 function M.setup(cfg)
    core.cfg = config.build(cfg)
@@ -201,11 +96,11 @@ function M.show()
 
 
    local buf = util.current_buf()
-   update(buf, false)
+   state.update(buf, false)
 
    for b, _ in pairs(core.crate_cache) do
       if b ~= buf then
-         update(b, false)
+         state.update(b, false)
       end
    end
 end
@@ -219,11 +114,11 @@ function M.toggle()
 end
 
 function M.update(buf)
-   update(buf, false)
+   state.update(buf, false)
 end
 
 function M.reload(buf)
-   update(buf, true)
+   state.update(buf, true)
 end
